[Working, but not clunky] - The issue has something to do with the `htmlTemplate()` function inheriting objects from the global environment.
This is talked about on the [Shiny Website](https://shiny.posit.co/r/articles/build/templates/) but I cannot figure it out with the use of `{{ }}`. I think part of the issue may be that in the example on the website that these `{{}}` sections to input R/Shiny into raw HTML, they only use these sections for actual Shiny inputs, while in my case **I am trying to insert UI and inputs, such that we have nested UI modules**. I think this is an issue because "to process the template’s R code, the htmlTemplate() function first creates a child environment of the global environment,  populates it with variables that were passed in as arguments to the function (like sliderValue), and then evaluates the template’s R code."

Either way, the rough and dirty solution just for now is just to combine all the scripts into one.